[ASV-1377] Extracted mopub ad listener to improve readability
